pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lonis"
version = "0.1.0"
description = "Bitmap-to-structured-design-data analyzer"
requires-python = ">=3.11"
dependencies = [
    "Pillow>=10.0",
    "opencv-python-headless>=4.8",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "scipy>=1.11",
]

[project.scripts]
lonis = "analyzer.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
